Finished bow, arrow and ProjectileManager
Reviewer: Jason

Weapon

DUN-77

Bow calculate the damage based on how long it is charged up to a maximum set time.
Base unit call the update of it's active weapon but only attack's when either the cooldown of the weapon is done or when "isHolding" is set to true.

BaseUnit has a ammo count that decreases when firing an arrow.

ProjectileManager updates all the arrows in the room.

Arrows won't leave the room and disappear when hitting a wall, door, object or ennemy.