Creation of dashing system
Filled the UseDash function from BaseUnit and added a dashTime/Timer, isDashing boolean and a dashing speed to modify the velocity of the rigidbody with the new speed and keep this speed until the timer has ended.

To fix : add a cooldown after the end of the dash because it is easy to just spam it. + Test